Remove deprecated method `argList`
The function `argList` was declared as deprecated before.
In addition it had an `@internal` TS doc comment which prevents the creation of the correct definition in the d.ts file. The missing definition caused an error, when exporting it further in the index.d.ts. Removing the function and the export resolves the compile errors.